fix: made edits from github comments
