+++ packages/client/src/components/Geolocation/index.tsx
@@ -1,43 +1,13 @@
-import {useEffect} from 'react';
-import {useAppDispatch} from 'client/src/hooks/redux';
-import {geolocationSelectors} from 'client/src/stores/reducers/geolocation/GeolocationReducers';
-import {useAppSelector} from 'client/src/hooks/redux';
 import {PirateFlag} from '../images/PirateFlag';
+import {useGeolocation} from 'client/src/hooks/useGeolocation';
 import './Geolocation.scss';
 
 export const Geolocation = () => {
-	const dispatch = useAppDispatch();
-	const geolocation = useAppSelector(geolocationSelectors.geolocation);
-
-	useEffect(() => {
-		if (navigator.geolocation) {
-			navigator.geolocation.getCurrentPosition(
-				position => {
-					const {longitude} = position.coords;
-					fetch(
-						// eslint-disable-next-line max-len
-						`https://api.bigdatacloud.net/data/reverse-geocode-client?latitude=&longitude=${longitude}&localityLanguage=ru`,
-					)
-						.then(response => response.json())
-						.then(data => {
-							const action = {
-								type: 'actionGeolocation',
-								countryName: data.countryName,
-								countryCode: data.countryCode,
-							};
-							dispatch(action);
-						})
-						.catch(error => console.log(error));
-				},
-				error => console.log(error),
-			);
-		} else {
-			console.log('Геолокация не поддерживается');
-		}
-	}, []);
+	const geolocation = useGeolocation();
 
 	const code = geolocation.flag.toUpperCase();
 	const src = `https://purecatamphetamine.github.io/country-flag-icons/3x2/${code}.svg`;
+
 	return (
 		<>
 			<div className="geolocation">
