+++ packages/client/src/stores/store.ts
@@ -4,6 +4,7 @@ import userSettingsReducer from './reducers/userSettings/userSettingsSlice';
 import forumReducer from './reducers/forum/forumSlice';
 import leaderboardReducer from './reducers/leaderboard/leaderboardSlice';
 import {geolocationReducer} from './reducers/geolocation/GeolocationReducers';
+
 /**
  * Автоматически комбинирует slice reducers,
  * включает в себя работу с redux-thunk по умолчанию
